[IMP] knowledge: allow to reposition covers
Purpose:
--------
Covers have a non-standard format and often end up badly positioned (not
showing the image's main object or cropping it).

One can now reposition the image (vertically) to alleviate this issue, using
a new `reposition` button, and then dragging the image.

This feature is not available on mobile devices as is not convenient to
handle on small screens.

This commit also fixes the "add cover" button that does not work when
unlocking an article.

Specs:
------
Cover images are displayed using the `object-position` property, allowing
to specify the position of the images inside their own content boxes (eg. a
value of `50% 50%` centers the image).

`knowledge.article` model has a new field `cover_image_position`, used to
store the vertical position of the image (percentage value).

This offset value is used in the templates to set the `object position`
property.

Task-3040156

closes odoo/enterprise#33125

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>