[FIX] delivery: sendcloud labels traceback
This PR addresses a bug concerning a traceback when validating a sale order with sendcloud as carrier The traceback indicated that the attachements that will be send in the chatter must be passed as an array of ids. Therefore the fix only changes the ir_document objects to a list of their ids.

To recreate the bug =>
- Create and configure Sendcloud as a delivery carrier.
- Use Sendcloud as a delivery carrier for a sale from the website
- Once payed, the order's delivery must be validated => traceback

closes odoo/enterprise#36261

Signed-off-by: William Henrotin (whe) <whe@odoo.com>