[IMP] account_reports: Aged Payable/Aged Receivable: speed-up unfold_all perf
On both these reports, when a large amount of data was to be displayed, unfold_all filter timed out. Indeed, all the partner lines were unfolded one by one, each generating a new SQL query. We optimize this by introducing the possibility for custom engines to use a custom_unfold_all_batch_data_generator (already used for this kind of situation on reports with dynamic lines), and using it on both reports.

Note that when the report contains too many lines to display, the browser still might not support displaying it (Chrome seems to perform quite poorly on that side). For this reason, since such huge reports will not be accessible fully unfolded from the UI as-is, we partially revert https://github.com/odoo/enterprise/commit/5787cd561028a8f5f0882c99316c429dfadfab10 in order to reintroduce the fact printing and Aged Payable or Aged Receivable report will by default unfold all the lines of the report. This way, printing the report to xlsx will directly provide all data required for an audit of the report, and the user won't have to enable the unfold_all filter manually in the UI.

closes odoo/enterprise#36367

X-original-commit: 1a255be857c1449063f1ee53717fd7e388b9f5ad
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>