[IMP] spreadsheet_edition: disable inserting in spreadsheet when pivot groupbys have duplicates
This commit disables INSERT IN SPREADSHEET button when users want to insert pivot table with duplicated groupbys in either rows or cols, or cols/rows have the same groupbys into spreadsheet. The tooltip shows the reason.

Currently when the pivot table with duplicate groupbys only in cols/rows, or rows and cols have same groupbys, in spreadsheet the values are gone. Since these cases are meaningless in most of time, it is more economical to just prevent this kind of insertion than fixing the bug, which probably needs a big change in the way we get the pivot values.

task 3068786

closes odoo/enterprise#36412

X-original-commit: 1509061f2208d2be68ce066e27b3b331a69f8af2
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>
Signed-off-by: Yang Chenyun (chya) <chya@odoo.com>