[IMP] documents: improve document request process
This commit is a review of the document request process.
The owner of the file will now be the requester, not the requestee.
This has an impact on the activity view as the requester will be
displayed next to the activity instead of the requestee.

A reminder mail template has been added to be able to manually send a
reminder on late document requests.

TaskId-3107229

closes odoo/enterprise#37619

Signed-off-by: Stéphane Debauche (std) <std@odoo.com>