[IMP] documents_accounting: switching journals on workflow.rule
Some customers use more than one journal to input accounting, ie:
one journal for expense or visa and one for "standard" purchase
or 2 banks journals.

We want to let them decide which journal will be used in their workflow.

This commit adds the selection of journal in the action form. By default,
the first one of the list is selected to keep the actual behavior, but
allow the user to select another one.

[community](https://github.com/odoo/odoo/pull/111328)
[enterprise](https://github.com/odoo/enterprise/pull/36308)

closes odoo/enterprise#36308

Related: odoo/odoo#111328
Signed-off-by: William André (wan) <wan@odoo.com>