[IMP] industry_fsm{,_report,_sale}: hide customer for non fsm task/project
Before this commit user able to see and add customers on non fsm and
non billable tasks/projects which is not provide any significance as
user don't need customer for normal tasks/projects.

So, in this commit hide customer field and also move partner phone
and city to field service app as it was only usefull for FSM project
and task.

task-3141350

closes odoo/enterprise#36459

Related: odoo/upgrade#4274
Related: odoo/odoo#111335
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>