[FIX] sign: render well arabic inputs
Steps to reproduce:
- Go to System Parameters, Add:
key: sign.use_custom_font
value: FreeSerifBold

(Tested with FreeSerif and Sun-ExtA as well).

-  Go to Sign > Go to any template, add Text fields > Sign Now.
-  Paste in any Arabic language segment, Complete signing.
-  The signed document in Odoo shows the Arabic correctly > Download it >

Issue:
The Arabic is flipped and changed to be incorrect in downloaded document.

Cause:
The reportlab library don't render well arabic words and always use the Left-to-right writing

Solution:
Use of arabic-reshaper and python3-bidi to render it well but it's possible for now as the actual
version of Ubuntu and Debian in Saas and Sh can handle so it will used as permanent fix in the future.
As a hot fix we just reverse the text when it's arabic. The form isn't the best one(ligatures between characters), but the meaning stays and the text can be read.

opw-3111259

closes odoo/enterprise#42397

X-original-commit: 05852818c5f86ff35d72a53cdcc260bfad443991
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Djoumatchoua Eteil Junior (etdj) <etdj@odoo.com>