[FIX] web_map: Improve mockRead performance
A commit in https://github.com/odoo/odoo/pull/110819 improves the
performance of mockRead which was too slow with a large number of
records. The way mockRead was rewritten broke some tests in
map_view_mobile_tests because the model "res.partner" was not defined
in serverData. This commit adds it to fix those tests.

Part-of: odoo/enterprise#36202
Co-authored-by: Bruno Boi <boi@odoo.com>
Co-authored-by: Mathieu Duckerts-Antoine <dam@odoo.com>