[FIX] (website_)appointment: restore controller method override
The overridden `_get_customer_country` method could not be called in
two places where we referenced the base appointment controller
instead of:
* in `Appointment` itself, the current class which could have been
 extended
* for `_search_get_detail`, the `WebsiteAppointment` controller

We here restore this override lost with 6ca7f201.

The test included fails without both fix parts of this commit.

An extra-request is therefore necessary for the visitor lookup.

Task-3151448

closes odoo/enterprise#37730

X-original-commit: 1b64e9e1399c30a031c71753628a972c6069415b
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>