[IMP] account_accountant: Bank Statement attachment
When importing a statement, the printed PDF is attached to the statement, and is accessible in the chatter of all the journal entries related to that statement.
We don't have this feature when statements are created manually.
With this PR, a new print action is added to the statement views. When clicked, the PDF is generated, attached to the statement record, and downloaded. The preview is then available on the relevant Journal Entries.

This action replaces the original report action (see odoo/odoo#111133).

Task-3151040

closes odoo/enterprise#36355

Signed-off-by: Laurent Smet <las@odoo.com>