[IMP] account_accountant: Bank Statement attachment
When importing a statement, the printed PDF is attached to the statement, and is accessible in the chatter of all the journal entries related to that statement.
We don't have this feature when statements are created manually.
With this PR, a new print action is added to the statement views. When clicked, the PDF is generated, attached to the statement record, and downloaded. The preview is then available on the relevant Journal Entries.

This action replaces the original report action (see odoo/odoo#111133).

Task-3151040

closes odoo/enterprise#36559

X-original-commit: 8cdd39f8548b7535ebfca4a454ae31ac1a0fef6c
Related: odoo/odoo#111542
Signed-off-by: Laurent Smet <las@odoo.com>
Signed-off-by: Ayob Habib (ayh) <ayh@odoo.com>