[FIX] mrp_plm: correctly calc latest_bom_id
When the user clicks on the "Update BoM" button, in some cases, the BoM
isn't updated.

How to reproduce:
- Create a BoM for a product template (not fill the product variant);
- Create a manufacturing order using this BoM;
- In PLM, create a new version of this BoM with an ECO and apply the
  changes -> The previously created BoM should be archived now.
- Go back on the MO, the button "Update BoM" should be visible next to
  the Bill of Material fields;
- Click on it -> The BoM isn't updated, the old (archive) version of the
  BoM is still used by the MO.

The issue happened in the `_get_active_version` BoM's method.
This method used a domain based on the `product_id` so it didn't find
the BoM's new version if they use the `product_tmpl_id` field instead.

Also, makes some change in the `mrp.production` `_compute_latest_bom_id`
to make it clearer (it seems counterintuitive to assign MO's `bom_id` as
the MO's `latest_bom_id`).

Part-of: odoo/enterprise#35969