[FIX] account_reports: avoid regrouping the errors and the non-errors…
… in the Accounts Coverage Report

Currently the trie regrouping feature of the Accounts Coverage Report
regroups both errors and non-errors accounts. This is problematic as we
can see on the following two examples:

1. If we have the same error on the codes 4282, 4284, 4286 but no error
on the code 4287, these 3 codes will be regrouped under 428 even though
4287 is not flagged as an error because it is not in the trie, so now we
add also to the trie non-errors to be sure to avoid regrouping too much.

2. If we have an error on the code 201 (code reported multiple times on
the same line) but nowhere else in the up the trie (20, and 2), the
reported account code with an error will be "2" and not "201" because of
 this regrouping.

X-original-commit: ff3cd52e80da58a23236072a85f5e027f4ae1f4f
Part-of: odoo/enterprise#37784