[FIX] web_studio: inline view xpath with duplicate field
Steps to reproduce
==================

- Go to Manufacturing (with work orders enabled)
- Go to the form view of a manufacturing order
- Open studio
- Click on work orders
- Try to edit the list subview
-> Single sub-view arch not found for xpath: `/form[1]/sheet[1]/notebook[1]/page[2]/field[1]/tree`

Cause of the issue
==================

In order to edit an inline view if it doesn't exist, studio creates one.
The one created in this case has the following xpath `//field[@name='workorder_ids']`
This isn't enough to uniquely identify the proper node, as in this case,
there is another field with the same name before the one we want.

Solution
========

Send the full xpath to the backend. Also don't add the field as it is
already present in the full xpath

opw-3115867

closes odoo/enterprise#36939

X-original-commit: 7ab37117f9813fd13997ba8ac547990d1c4ccac8
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>
Signed-off-by: Hubert Van De Walle <huvw@odoo.com>