[FIX] l10n_{co,mx,pe}_edi: add groups to username/password fields
Often we have one account for connecting to a web service (e.g. sending invoices to the gov) and as it can be different per company. The credentials fields are stored on the company but there is no security groups=`'base.group_system'` , so these are visible by every user, because `res_company` fields can be seen by everyone. 
This commits adds `groups='base.group_system'` on credentials fields. This also requires a `sudo()` whenever we call these fields.

task-3044048

closes odoo/enterprise#36487

X-original-commit: 26c8d33e321625261288412893f74b3311306edc
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Aliya Tastemirova (alta) <alta@odoo.com>