[FIX] stock_barcode: test_gs1_receipt_lot_serial
Following this fix: https://github.com/odoo/enterprise/pull/33866
In BarcodeModel `_findLine` method, we check the line `picking_id` to
know if we are in a picking or in an inventory adjustment.
The issue is we don't fetch the move lines picking id in `stock_barcode`
but we do once `stock_barcode_picking_batch` is installed.
To fix that, we check explicitly if the current model is `stock.quant`.

Also, in the test `test_gs1_receipt_lot_serial`, actives the tracking
group because this test checks lines tracked by lots.

runbot build error: 15279

closes odoo/enterprise#37831

X-original-commit: dc4dfba9f1762f3632dc943c1161e75d1a79cab5
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Steve Van Essche <svs@odoo.com>