[FIX] account_avatax: allow duplicate document codes in sandbox mode
Avatax stopped providing free trials. Instead users should now use a
shared partner sandbox account. The sandbox account will have a single
company and in that company the document codes for the invoices (field
`avatax_unique_code`, e.g. Journal Entry 1) must be unique. If an
invoice with the same document code is submitted Avatax will return an
"Invalid DocStatus" error. The `avatax_unique_code` is created using
the database ID, so across different Odoo instances there will be
collisions.

In order to work around this Avatax recommend we switch to the
"createoradjust" endpoint [1]. It's a drop-in replacement for "create"
that will updated even committed invoices. It only errors when an
invoice is reported to a tax authority, but that will never happen in
a sandbox environment.

This commit uses that endpoint in sandbox mode. For production nothing
changes.

[1] https://developer.avalara.com/api-reference/avatax/rest/v2/methods/Transactions/CreateOrAdjustTransaction

opw-3152188

closes odoo/enterprise#36541

X-original-commit: 72bff0f80e830697fb98df3b095e586fbe5493f4
Signed-off-by: William André (wan) <wan@odoo.com>