[IMP] l10n_mx_edi_40: partner fiscal regime setup
The Mexican localization has a custom field for the fiscal regime.
Currently this field is only shown for Mexican partners and defaulting to `601` (General de Ley Personas Morales).
Foreign partners can legally only have the regime `616` (Sin obligaciones fiscales).
Since we default to `601` and the field is not shown for foreign partners, the field is always incorrect.

This change computes the field depending on whether the partner is Mexican or not and also allows the user to change it manually.
The field is shown for all partners now, so foreign partners can get the correct fiscal regime.

task-3143841

closes odoo/enterprise#36386

Related: odoo/upgrade#4277
Related: odoo/odoo#111230
Signed-off-by: Quentin De Paoli <qdp@odoo.com>