[FIX] sale_renting: discount remain on SOL after picking
Steps to reproduce:
- Activate the discount in the Sales settings;
- Create a SO with a product that can be rented;
- Add a discount and confirm the order;
- Validate the pickup.

Current behaviour:
The discount disappears on the SO line.

Expected behaviour:
The disount remains on the SO line.

This commit prevents updating the `product_uom_qty` of the line if the
delivered quantity does not become larger than set on the SO line. The
reason is that a change in the `product_uom_qty` triggers a compute of
the SOL price that does not take the discount into account.

OPW-3146922

closes odoo/enterprise#36421

X-original-commit: f14ef1413be4b2eb43a0bb5384c4451be73fa7ab
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>
Signed-off-by: Vallaeys Valentin (vava) <vava@odoo.com>