[FIX] appointment: fix 'anytime' appointment type duplication detection
An appointment.type of category 'anytime' can only exist once for a given
res.users.

The code trying to detect and raise an error upon finding a duplicate record
was not working and raised on invalid cases.
Existing tests only worked because of side-effects.

The code was fixed and extra use cases are now tested to ensure the behavior.

Task-3159839

closes odoo/enterprise#36427

X-original-commit: 025ae300c91d5962aff6d8394e54f4bc5913757e
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>