[FIX] planning: alloc. hours aggr. doesn't depend on company cal.
Current behaviour:
The aggregated allocated hours are based on company calendar when
the associated slots don't have an attached resource.

Expected behaviour:
The aggregated allocated hours shouldn't be based on company
calendar when there is no resource to the assigned slot.
(To follow current behaviour of the computation of individual slots'
allocated hours)

Steps to reproduce:
- Install Planning
- Create a new slot from 10AM -> 18:30PM
- The allocated hours on the slot is 8h (because when the slot
  doesn't have resource, we take the min between the diff of
  `end_time` and `start_time` and the `hours_per_day` of the company's
  calendar)
- But the aggregated allocation for the day is 6h (instead of 8h),
  because it takes the company calendar into consideration, which
  has worktime internals 10->12 + 13->17 (6h in total).

Reason for the problem:
The front-end fallbacks to the company's working calendar when there
are no resources on the slots.

Fix:
When computing the aggregate, if the resource is empty, we use the
same formula as the backend to compute the allocation_hours for the day.
We add an rpc call to retrieve the company's hours_per_day.

Affected versions:
- 16.0
- saas-16.1
- master

opw-3134643

closes odoo/enterprise#36302

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>