[FIX] knowledge: memorise graph type and measure when inserting embedded view
When the user opens a graph view, changes the measure, changes the type
of the graph and inserts the graph in an article of Knowledge, the system
completely omit the graph type and the measure set by the user. It simply
loads the graph with the default settings.

With this commit, the embedded views will now load by default the graph
that was active when inserting the view. This new behavior will allow
the user to insert the graph they want.

task-3056268

X-original-commit: fd345719b737520fe015aab3c4fe2e17d7ae9ff4
Part-of: odoo/enterprise#37046