[FIX] knowledge: memorise list order when inserting embedded view
When the user opens a list view, orders the records by clicking on a
column and inserts the list view in an article of Knowledge, the system
omits the sorting order set by the user.

With this commit, the embedded views will now keep the sorting order set
when inserting the view. This new behavior will allow the user to customize
the embedded view even more.

task-3056268

closes odoo/enterprise#33707

Related: odoo/odoo#105167
Signed-off-by: David Beguin (dbe) <dbe@odoo.com>