[FIX] web_studio: `<t groups="...">...</t>` must always remain
This revision is related to odoo/enterprise#34752

When a `<t>` node with a `groups` attribute is set in a view,
this `<t>` block must remain, whatever the case,
so the view sent to the web client with Studio represents
the actual view in the backend.
This is so the xpath computed by Studio is based on the same blocks
in the actual view in the database then what is sent by `get_views`.

It was already the case when the user did not have the group,
but it wasn't the case when the user had the group.
In such a case the call to super was done with the `groups` attribute
left, and the call to super therefore removed the `<t>` block itself
(moving its content out of it first).

This revision solves the below use case:

Steps to reproduce:

- Install Sales, CRM and studio
- Go to Sales > Orders > Sales Teams
- Now activate studio, and try to modify the progressbar widget.

Issue:

We are going to receive an error message and we will be unable to change
ths widget.

opw-3106068

closes odoo/odoo#111111

closes odoo/enterprise#36513

X-original-commit: 03d29caf2072fb7183317e81f1d00116c14a571e
Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>