[FIX] account_reports: analytic filter & groupby
With this commit, we ensure that the analytic
filter is applied on the analytic group by
columns.

Steps to reproduce:

- Activate analytic accounting
- Activate analytic filter on P&L
- Create a bill with two analytic accounts
  on the invoice line
- Go to P&L report
- Group by the two analytic accounts
- Filter by any of the analytics accounts
-> Issue: nothing is displayed in the groupby
   columns.

This is because we add twice the analytic account
as where parameter/clause in every cases, it should
be the case only if the account is not already
present in the query.

opw-3085092

closes odoo/enterprise#36618

Signed-off-by: William André (wan) <wan@odoo.com>