[FIX] spreadsheet: linking odoo menu not jumping to odoo home
This commit fix the problem when inserting a link to an odoo menu,
clicking the menu item in the dropdown will jump to odoo home page, and the insertion fails.

The root cause is the menu item uses `a` tag and has a `href` leading to home page.
The solution is to prevent the default clicking behavior.

This commit also fixs a small bug of showing currently selected menu item,
and make the menu selector show all apps as well.
In the past, menu link only shows available menus, not apps.
So when users get a menu link from an app,
 if they delete it accidentally, it cannot be recovered.

task 3159543

closes odoo/enterprise#36404

Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>