[REF] l10n_es_reports: pass period and year as arguments
If the "Trimester monthly report" option is enabled when generating a
BOE 349 tax report, it reduces `date_to` in the `options` dict by 1
month. If later we try to calculate the period and year from those
modified values, that will fail in the trimester case because
`_get_mod_period_and_year` is expecting 3 months instead of 2.

To avoid this, we pass the period and year into
`_write_type2_refund_records` as arguments so the values don't have to
be recalculated again.

This wasn't done in the original bug fix in
https://github.com/odoo/enterprise/pull/36228 to avoid changing the
function signature.

opw-3131030

closes odoo/enterprise#36550

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>