[FIX] account_reports, sale_subscription_dashboard: escape JSON error when downloading report
__Description of the issue:__

When something goes wrong while downloading a report file, a 500 error
is sent as JSON. However the frontend interprets this response as HTML
and then try to parse the text content as JSON.

Most of the time this works, but if the response contains any HTML tags,
like `<lambda>` from a Python stacktrace, the JSON response will get
misinterpreted as HTML instead of regular text, causing the subsequent
JSON interpretation to fail.

The end result for the user is that empty tracebacks will be displayed
instead of User Errors or actual tracebacks.

__Desired behavior:__

The JSON response is HTML escaped before being sent and will therefore
be correctly parsed and displayed to the user.

It basically restore what was done in #33443

Community: odoo/odoo#111433
See also: #35308

opw-3131030
opw-3085277

closes odoo/enterprise#36978

X-original-commit: b1c97265ff86773ae0e11b760c919815b53165ae
Related: odoo/odoo#112453
Signed-off-by: Julien Castiaux (juc) <juc@odoo.com>