[FIX] account_trial_balance_report: order column
Before this commit it was not possible to sort lines on the trial balance.

The problem was due to the fact that we use a _custom_options_initializer that adds columns (initial and end balance) **after** we order the columns. At the moment we do the order, we don't have all the columns yet and since we check the index of the ordered column, it doesn't work. (For example, if we want to order column at index 3 but we only have 2 columns, the index will never match)

To fix that, we compute (again) the _init_options_order_column **after** we add the columns.

closes odoo/enterprise#36542

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>