[FIX] sale_subscription_dashboard: fix format monetary
Before this commit, default digits were given to the
monetary formatter but those were not used as we also
gave a currency. since commit https://github.com/odoo-dev/odoo/commit/996cc74745a2fb72547053991a79d233d6916587,
the digits are used if given to the formatter and caused
a formatting error as the digits were wrong.
This commit removes the given digits to use the currency ones.

closes odoo/enterprise#36568

Related: odoo/odoo#111308
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>