[FIX] data_merge: batch cleanup cron and speedup cron
The cleanup cron would sometimes timeout on databases with a large
number of records and the changes would not be commited to the database,
hence creating an infinite loop of failing crons.

The cleanup operations are now performed on small batches and the
changes commited for each batch to ensure smooth operations.

The performance of the cron has been improved as well, before all the
original records were retrieved one by one, instead of being retrieved
in batch (per group).

task-2977145

closes odoo/enterprise#36718

X-original-commit: 4ee09ec17fffbb2e6ec3aeb98c25aed9a1e14dd6
Signed-off-by: Kevin Baptiste <kba@odoo.com>