[FIX] knowledge: fix undefined messaging variable in renderer
Currently, the `useEffect` function uses a variable that will be set on
the instance of the Knowledge renderer when the messaging service has
finished loading. The loading occurs asynchronously.

Sometimes, it may happen that the Knowledge renderer loads faster than
the messaging service. In that case, the message service reference used
in the `useEffect` functions of the Knowledge renderer can be undefined.
The script can thus raise an error when it calls a function on the
message service reference.

To fix the issue, we will move the code of the `useEffect` function using
the messaging service to the promise loading the message service. The
promise will have the responsibility to (1) load the service, (2) set
the `messaging` variable used by the Knowledge renderer and (3) add some
event listeners to the messaging bus. Those listeners will be removed
when the Knowledge renderer is destroyed. With this approach, we do not
have to wait for the messaging service to be loaded to start using Knowledge.

task-3164403

closes odoo/enterprise#36554

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>