[FIX] account_report: sortable
Change the way options['order_column'] works. Instead of just having the index of the column to be sort (with potentially a - in front) now it is a dictionary with the expression_label of the column to be sort and the direction of the sort (ASC or DESC).

This fixes the issue were if we have columns not displayed because they are empty it will not break the sort since it is not based on an index (which can change) but on an expression label (which can **not** change).

closes odoo/enterprise#36954

X-original-commit: 7bb7c29abe02e497934ff42a85a0d0f6721ee172
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: rugo-odoo <rugo@odoo.com>