[FIX] account_asset: fix import of asset models
Currently, when asset/deferred revenue/deferred expense models are imported, an asset/deferred revenue/deferred expense is created instead of a model. This happens because the model creation depends on state, which gets overwritten in `create` starting from 16.0.
State should not be overwritten if `default_state` is present in the context.

opw-3140284

closes odoo/enterprise#36583

Signed-off-by: William André (wan) <wan@odoo.com>