[FIX] stock_barcode_mrp: scan kit with tracked components
If a user scans a kit that has a tracked component, he will not be
able to validate the picking

To reproduce the issue:
1. Create a kit with two storable products P_kit and P_compo
   - P_kit:
     - Barcode: kit
   - P_compo:
     - Tracking: Lot
2. Barcode > Operations > Receipt, Create
3. Scan kit
4. Validate

Error: a user error is raised "You need to supply a Lot/Serial
number for products [...]." The user is now stuck as he does not
have the possibility to edit the SML of the component

When validating, we confirm the SM of the kit, so it will be
exploded into a SM for the component. Then, it will try to call
`_action_done` on that SM but since we didn't provide any lot, it
will raise the user error.

OPW-3015933

closes odoo/enterprise#36682

X-original-commit: 5ddb470bd1ad5dad244a36d800d93349518add4c
Related: odoo/odoo#111881
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Adrien Widart <awt@odoo.com>