[FIX] data_merge: don't update default data
If we keep the data in noupdate=0 each time we update the module (like
in an upgrade 14->15 for example) we recompute many fields due to the
write done by the load.

It has been observed that even if the write doesn't change anything in
the DB the computation could cause MemoryError or maximum recursion
depth exceed.

closes odoo/enterprise#37447

X-original-commit: d3cf2e768b2d7c97fd39ca7121a83d1d91354dd3
Signed-off-by: Kevin Baptiste <kba@odoo.com>