[FIX] data_merge: don't update default data
If we keep the data in noupdate=0 each time we update the module (like
in an upgrade 14->15 for example) we recompute many fields due to the
write done by the load.

It has been observed that even if the write doesn't change anything in
the DB the computation could cause MemoryError or maximum recursion
depth exceed.

closes odoo/enterprise#37401

X-original-commit: 0093580b19e33abcd78adcd0971e3ed90dde767e
Signed-off-by: Kevin Baptiste <kba@odoo.com>