[IMP] sale_subscription : No invoice after the end_date
This PR disallow the user to invoice a customer for a period after
the end_date.
The user will be warned and invited to change the end_date or renew
the contract instead.

There is no change for automatic invoicing as the CRON should close
ending contract beforehand.

task-id : 3151693

closes odoo/enterprise#36611

Related: odoo/upgrade#4293
Signed-off-by: Arnaud Joset <arj@odoo.com>