[IMP] account_reports: view improvements
This PR allow to do some UI improvements in the reports so that they are more readable for users.
This PR adds the possibility to have highlight on hover (was present in v15 but no more in v16), also change some coloring (lines and subtable) on the journal report to have a more readable report.
Finally the date present was repeated in the filter and on top of the report. When having no comparison, the date on top of the report is not displayed anymore to avoid repetition.

closes odoo/enterprise#37147

Task-id: 3166902
X-original-commit: f33469f81b0f3eccb89873c85532f34054f82bb6
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>