[FIX] documents_hr: correct message when clicking Documents stat button
When the user clicked the Documents stat button
from the Employee's form view, they would get a
ValidationError in case the Employee had no
address_home_id. However the message of the
ValidationError did not specify that a private
address is needed (not the address of the company)
making it confusing for the user.

opw-3144207

closes odoo/enterprise#36651

Note: the same behavior can be reproduced in 16.0, so we would need to forward-port this branch.
X-original-commit: e96682799bc0d5468693df2ff9c1d569d9a875c4
Signed-off-by: Kevin Baptiste <kba@odoo.com>