[IMP] mrp_plm: enable update BoM in mrp
The "Update BoM" action button will now be displayed also when there is
a difference between the BoM's lines and the MO's components, so a part
of the related code was moved in `mrp`.
Also, this action can now be called when the MO is confirmed, so the
widget to display the message about that when the MO is confirmed is now
obsolete and was removed.

See odoo/odoo#110227
task-3089368

Part-of: odoo/enterprise#35969