[FIX] appointment: show name of single user when sharing random appointment
Before, the name did not show in title because the appointment
type is in 'random'. It only showed at the end of the process,
once the appointment was confirmed. This is not appropriate
when sharing a link for a single user. Indeed, the user will
never be sure its link works since its name will not appear
unless they take an appointment.

Therefore, we allow showing the name of the user for random
appointment types even if the appointment is in random assign
mode. There is no real randomness for single-user links.

We now show the name in title, as well as the user's details card
in the 'Details' screen, and their picture if the option is set in
backend.

Task-3165885

closes odoo/enterprise#36674

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>