[FIX] account_reports: correct computation of cross_report expressions referring to the same line with different date_scope
When using two cross_report expressions calling the same expression from another report with different date_scope values (typically, a Balance Sheet calling the Profit and Loss from this year and the previous year), the computed amount was wrong: both lines always were evaluated to the same value. This was because the grouping made on aggregations to get their result was totally disregarding the date_scope, and stored the result of both expressions under the same key.

X-original-commit: c561c5824b1a979210f77580f078f5091eb0ccf4
Part-of: odoo/enterprise#36841