[IMP] mrp: employee login in workorder kanban view
The goal of this commit is to adapt all log in part done in 6c856b82ef134cd2e1c0e4b4c579a43d66e25d3d to the kanban view.
There is a lot of common code between the kanban and list controller so employeeHook will regroup the common functions/behaviors.

In addition, the workorder wizard helping to track the time spend on the workorders has been automated.
The name of the employee will automatically be the name of the admin of the session.
The productivity will also be updated based on the duration
The duration, start date and end date will update based on the two other ones.
This changes will ease the addition of time trackings.

related : https://github.com/odoo/odoo/pull/107473

Part-of: odoo/enterprise#34786