[FIX] industry_fsm_stock: reserve lines with fsm lot
If, from a task, we add several tracked products, the reservation of
the associated delivery will not work.

To reproduce the issue:
(Use demo data)
1. Create a product P:
   - Type: Storable
   - Tracking: USN
2. Update the on hand qty of P:
   - 1 x P with USN01
   - 1 x P with USN02
3. Create and confirm a SO with
   - 1 x Field Service
4. Open the related task
5. Open the products
6. Add P:
   - 1 x USN01
   - 1 x USN02
7. Back to the SO, open the delivery

Error: Only USN01 has been reserved. If the user clicks on "Check
availability", nothing changes. Suppose the delivery also contains
some other tracked products (added at step 6), none will be reserved

When checking the availability, for each SM, we try to reserve a
quantity with a specific lot. To get this lot, we get the associated
procurement group (which is useful in case of multistep delivery),
its stock moves (still ok) and, among all the associated SOL, we
arbitrarily select the first one (not ok).

In the above case, it means that we always take the SOL with USN01,
even when trying to reserve the quantity of the second SM (i.e., the
one related to USN02). As a result, it will only works the first
time. Then, since USN01 is already assigned, the reservation will
not do anything (this is the reason why none of the others lines is
reserved)

OPW-3146832

closes odoo/enterprise#36892

X-original-commit: d09531c11a590492fb4d45e6fea599f2444e3e63
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Adrien Widart <awt@odoo.com>