[FIX] knowledge: updateBehaviors in edit mode only if odooEditor
How to reproduce:
- open an article in Knowledge
- create a table of contents (/toc) or an outline (/outline)
- lock the article
- unlock the article

Current behavior:
traceback

Expected behavior:
no traceback

Fix:
- `startWysiwyg` method from `html_field` is an asynchronous method that is
sometimes called synchronously (`onMounted`).
- Behavior Components are instanciated in independant Owl Apps inside the
  `html_field`, by the `updateBehaviors` method.
Therefore we have to make sure that when the html_field is in edit mode, we only
call `updateBehaviors` when the `wysiwyg` is ready
When switching from locked mode (readonly) to unlocked mode (edit), the
html_field undergoes `onPatched`, but the wysiwyg is not ready yet (with the
odooEditor), so we can't `updateBehaviors` at that time. It will be done in the
`startWysiwyg` method instead, after properly awaiting the OdooEditor.

Task-3165912

closes odoo/enterprise#36590

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>