[FIX] knowledge: cancel restore access desynced child article
This commit fixes an issue where, when you would cancel the confirmation dialog
when trying to restore the access to a children article a crash would be thrown
at the user.

This is caused by the discard function not being properly instanciated when
no function is given as a parameter of the `_showConfirmDialog` function.
Now instead of giving an unbound function as a discard method we give the same
function as before but bound to the PermissionPanel object.

task-3171926

closes odoo/enterprise#36977

X-original-commit: f4eb876346f7527108bd6ee50c628322b5a2e185
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>