[IMP] account_asset: make board invisible when asset modified
Before this commit, we could:
- Create an asset
- Compute the depreciation board
- Modify some values like the original value, or the accounts, ...
- Confirm the asset
=> the depreciation board is not recomputed, so the values are inconsistent

We will now change the values of the accounts and journal of
the entries when changed on the asset.
The reason being that when the user wants to see the values of the entries,
we write on the asset, so he will see the right values.

For other fields (the values, dates or type of computation),
we will unlink the depreciation board.
When modifying this field, the user will see that his
depreciation board has disappeared, and won't be surprised.
They can still undo their changes, to have their old depreciation board.

task-2937679

closes odoo/enterprise#37255

X-original-commit: d5a3126a4ae6cc767b0469eb82477fb14cbe8809
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>