[REF] web_grid: convert the grid view in OWL
The grid view structure is now built in the JS as the others views based on the data fetched.
Those data are now fetched via read_group method as the others views to be able to use the
sample server without reimplementing the python methods implemented only for the grid in JS.

It means the `read_grid` and `read_grid_grouped` are no longer used and removed.
The group expand was not correctly supported in the grid view, only the section field
triggered the group_expand of the field used for section and if the group expand is needed
for the fields used in rows then the inner method called by `read_grip` and
`read_grid_grouped` is override. Currently, the group expand will be prepared in JS to call
an extra read_group to get those data.

The advantage by doing that is the data passed has been drastically reduced since the
structure of the grid view (header, rows, cells) are no longer computed in server side.

- `adjust_grid` method has been replaced by `grid_update_cell`
- `grid_unavailability` method has been added to be able to return the unavailability dates
  of the current user or the employees in the grid section (in the case of timesheets).

task-2702636

Co-authored-by: Xavier BOL <xbo@odoo.com>
Co-authored-by: Elisabeth Dickinson <edi@odoo.com>