[REF] account_consolidation: adapt consolidation grid view in OWL
This commit converts the custom grid view for `consolidation.journal.line`
model in OWL. Also, the override made to make cell in readonly if the
journal is auto-generated (`auto_generated=True`) is now done in JS
during the generation of the columns. More precisely, `auto_generated`
is also fetched in the data (the journals) fetched to build grid columns
to get the readonly columns.

Co-authored-by: Laurent Stukkens (LTU) <ltu@odoo.com>