[REF] project_timesheet_forecast: adapt group expand based on shift
Before this commit, when the grid view was in legacy, the group expand
based on shifts are made in python when `read_grid` and
`read_grid_grouped` are called.

This commit adapts the code to do it in JS after the main `read_group`
call to be able to call another `read_group` to fetch the additional
data based on shifts.

task-2702636